remove paragraph id sort
This sort was not working as originally intended. It was sort in lexographical order instead of in ascending by number. This was causing weird paragraph orders such as verse 1, verse 10, verse 2 instead of verse 1, verse 2, verse 10.

Rather than fix so it orders by numerical value, I decided to simply remove the sorting and allow the url order to determine the display order. This simplifies things and also allows for having verses display in a different order than ascending such as verse 11 then verse 2, which sometimes Come Follow Me does.