Come follow me task generator (#79)
Introduces the ability to enable "experimental features" and it adds a new experimental feature for generating study tasks from a URL. This is an early idea for my personal study that I wanted available for me to use. I will likely continue to evolve it and bring it out of the experimental stage, but I haven't quite figured out the future of it.